pyproject.toml
[project]
name = "rag-solution"
version = "0.1.0"
description = "RAG Solution homework for 21.co"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "pymilvus[model]>=2.5.14",
    "sentence-transformers>=5.0.0",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

pythonpath = [
  "."
]
